requirements.txt
django==2.2
djangorestframework==3.9.2
django-ckeditor
django-import-export